Move filtering and tech-centred functions to hammer_tech
HammerTool has accumulated these functions but since they primarily only
reference the technology, they can move to hammer_tech.

It also enables library users to perform technology-centered action
without having to spawn a tool instance.
