+++ src/hammer-vlsi/hammer_vlsi/hammer_vlsi_impl.py
@@ -90,40 +90,6 @@ class HammerVLSISettings:
             return True
 
 
-# Library filter containing a filtering function, identifier tag, and a
-# short human-readable description.
-class LibraryFilter(NamedTuple('LibraryFilter', [
-    ('tag', str),
-    ('description', str),
-    # Is the resulting string intended to be a file?
-    ('is_file', bool),
-    # Function to extract desired string(s) out of the library.
-    ('extraction_func', Callable[[hammer_tech.Library], List[str]]),
-    # Additional filter function to use to exclude possible libraries.
-    ('filter_func', Optional[Callable[[hammer_tech.Library], bool]]),
-    # Sort function to control the order in which outputs are listed
-    ('sort_func', Optional[Callable[[hammer_tech.Library], Union[Number, str, tuple]]]),
-    # List of functions to call on the list-level (the list of elements generated by func) before output and
-    # post-processing.
-    ('extra_post_filter_funcs', List[Callable[[List[str]], List[str]]])
-])):
-    __slots__ = ()
-
-    @staticmethod
-    def new(
-            tag: str, description: str, is_file: bool,
-            extraction_func: Callable[[hammer_tech.Library], List[str]],
-            filter_func: Optional[Callable[[hammer_tech.Library], bool]] = None,
-            sort_func: Optional[Callable[[hammer_tech.Library], Union[Number, str, tuple]]] = None,
-            extra_post_filter_funcs: List[Callable[[List[str]], List[str]]] = []) -> "LibraryFilter":
-        """Convenience "constructor" with some default arguments."""
-        return LibraryFilter(
-            tag, description, is_file,
-            extraction_func,
-            filter_func,
-            sort_func,
-            list(extra_post_filter_funcs)
-        )
 
 
 from .hammer_tool import HammerTool, HammerToolStep
@@ -890,11 +856,11 @@ class CadenceTool(HasSDCSupport, HammerTool):
         pre_filters = optional_map(corner, lambda c: [self.filter_for_mmmc(voltage=c.voltage,
                                                                            temp=c.temp)])  # type: Optional[List[Callable[[hammer_tech.Library],bool]]]
 
-        lib_args = self.read_libs([self.timing_lib_with_ecsm_filter], self.to_plain_item, pre_filters=pre_filters)
+        lib_args = self.technology.read_libs([hammer_tech.filters.timing_lib_with_ecsm_filter], self.to_plain_item, pre_filters=pre_filters)
         return " ".join(lib_args)
 
     def get_mmmc_qrc(self, corner: MMMCCorner) -> str:
-        lib_args = self.read_libs([self.qrc_tech_filter],self.to_plain_item, pre_filters=[
+        lib_args = self.technology.read_libs([hammer_tech.filters.qrc_tech_filter],self.to_plain_item, pre_filters=[
             self.filter_for_mmmc(voltage=corner.voltage, temp=corner.temp)])
         return " ".join(lib_args)
 
@@ -904,8 +870,8 @@ class CadenceTool(HasSDCSupport, HammerTool):
 
         :return: List of qrc tech files separated by spaces
         """
-        lib_args = self.read_libs([
-            self.qrc_tech_filter
+        lib_args = self.technology.read_libs([
+            hammer_tech.filters.qrc_tech_filter
         ], self.to_plain_item)
         return " ".join(lib_args)
